Outsource public key and decryption function
